# Optimize sen slope(), yielding identical results.