Commit 329da7de09246f648968b1477f96dbffca669e76 sorts the array of slopes and uses the sorted array to find slope, lcl, and ucl. This avoids sorting the same data multiple times.

Commit 3949249064686e6cdee374b90b18a0744b0c4fe6 changes the order of checking for NaNs and iterating through the obs array to create the array of all slopes. This avoids adding NaNs in the array of slopes just to remove them immediatelly afterwards.

The runtimes gains are significant. Here are the results on an i5-8400 @ 2.80GHz:

Using v1.1.0: 433.10 sec
After applying 329da7d "Avoid gratuitous sorting": 348.60 sec
After applying both patches: 248.31 sec


pytest log using v1.1.0:


============================================================== test session starts ===============================================================
platform linux -- Python 3.9.2, pytest-6.0.2, py-1.10.0, pluggy-0.13.0
rootdir: /home/konstant/projects/atmo/dataeng-mannkendall
collected 16 items                                                                                                                               

test/test_mk_hardcoded.py .                                                                                                                [  6%]
test/test_mk_main.py ....                                                                                                                  [ 31%]
test/test_mk_stats.py ...                                                                                                                  [ 50%]
test/test_mk_tools.py ......                                                                                                               [ 87%]
test/test_mk_white.py ..                                                                                                                   [100%]

================================================================ warnings summary ================================================================
test/test_mk_main.py::test_mk_temp_aggr_single
  /home/konstant/projects/atmo/dataeng-mannkendall/src/mannkendall/mk_main.py:371: UserWarning: The trends for the temporal aggregation are not homogenous.
    warnings.warn('The trends for the temporal aggregation are not homogenous.')

test/test_mk_white.py::test_nanprewhite_arok
test/test_mk_white.py::test_prewhite
  /home/konstant/projects/atmo/dataeng-mannkendall/src/mannkendall/mk_white.py:94: UserWarning: No statistically significant autocorrelation.
    warnings.warn('No statistically significant autocorrelation.')

-- Docs: https://docs.pytest.org/en/stable/warnings.html
=================================================== 16 passed, 3 warnings in 433.10s (0:07:13) ===================================================


pytest log after applying 329da7d:


============================================================== test session starts ===============================================================
platform linux -- Python 3.9.2, pytest-6.0.2, py-1.10.0, pluggy-0.13.0
rootdir: /home/konstant/projects/atmo/dataeng-mannkendall
collected 16 items                                                                                                                               

test/test_mk_hardcoded.py .                                                                                                                [  6%]
test/test_mk_main.py ....                                                                                                                  [ 31%]
test/test_mk_stats.py ...                                                                                                                  [ 50%]
test/test_mk_tools.py ......                                                                                                               [ 87%]
test/test_mk_white.py ..                                                                                                                   [100%]

================================================================ warnings summary ================================================================
test/test_mk_main.py::test_mk_temp_aggr_single
  /home/konstant/projects/atmo/dataeng-mannkendall/src/mannkendall/mk_main.py:371: UserWarning: The trends for the temporal aggregation are not homogenous.
    warnings.warn('The trends for the temporal aggregation are not homogenous.')

test/test_mk_white.py::test_nanprewhite_arok
test/test_mk_white.py::test_prewhite
  /home/konstant/projects/atmo/dataeng-mannkendall/src/mannkendall/mk_white.py:94: UserWarning: No statistically significant autocorrelation.
    warnings.warn('No statistically significant autocorrelation.')

-- Docs: https://docs.pytest.org/en/stable/warnings.html
=================================================== 16 passed, 3 warnings in 348.60s (0:05:48) ===================================================



pytest log after applying both:

============================================================== test session starts ===============================================================
platform linux -- Python 3.9.2, pytest-6.0.2, py-1.10.0, pluggy-0.13.0
rootdir: /home/konstant/projects/atmo/dataeng-mannkendall
collected 16 items                                                                                                                               

test/test_mk_hardcoded.py .                                                                                                                [  6%]
test/test_mk_main.py ....                                                                                                                  [ 31%]
test/test_mk_stats.py ...                                                                                                                  [ 50%]
test/test_mk_tools.py ......                                                                                                               [ 87%]
test/test_mk_white.py ..                                                                                                                   [100%]

================================================================ warnings summary ================================================================
test/test_mk_main.py::test_mk_temp_aggr_single
  /home/konstant/projects/atmo/dataeng-mannkendall/src/mannkendall/mk_main.py:371: UserWarning: The trends for the temporal aggregation are not homogenous.
    warnings.warn('The trends for the temporal aggregation are not homogenous.')

test/test_mk_white.py::test_nanprewhite_arok
test/test_mk_white.py::test_prewhite
  /home/konstant/projects/atmo/dataeng-mannkendall/src/mannkendall/mk_white.py:94: UserWarning: No statistically significant autocorrelation.
    warnings.warn('No statistically significant autocorrelation.')

-- Docs: https://docs.pytest.org/en/stable/warnings.html
=================================================== 16 passed, 3 warnings in 248.31s (0:04:08) ===================================================