@fpavogt 2022-07-07
Hey @stasinos, thanks a lot for this PR ! The computational gain by not sorting everywhere is indeed significant. Since I'm currentlz preparing PR #34 to fix the docs (etc...), I'll point your PR to the `develop_vof` branch, and I'll then merge it all inside `develop` down the line.

@mcollaudcoen, I would be in favor of merging this contribution. But this would imply the existence of a first (very small, but still) code mismatch between the different implementations. Still, at roughly 50% speed gain, these changes have a very important effect, and no impact on the results themselves.