@fpavogt 2022-07-07
Thanks for the prompt update of the CHANGELOG ... and sorry about the resulting conflict. Feel free to simply undo that change, and I'll add your entry to the CHANGELOG myself.